ProyectBE/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.3
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
